feat: Add converter function to convert numbers to a given base
This commit adds a new function called converter to the 10-converter.js file. The function takes in a base as an argument and returns a convert function. The convert function takes in a number and converts it to a string representation in the given base. This functionality will be useful for converting numbers to different bases.