feat: Add code to concatenate two files
This commit adds code to the 102-concat.js file that reads two files and concatenates their contents into a new file. The first file is read using the fs.readFileSync method and converted to a string. The same is done for the second file. The contents of the first and second files are then concatenated using the + operator. Finally, the concatenated contents are written to a new file using the fs.writeFileSync method. This functionality will be useful for combining the contents of multiple files.