feat: Add script to connect to server with private key
This commit adds a script to the 0x0B-ssh/0-use_a_private_key file that connects to the server 100.25.221.149 using the private key located at ~/.ssh/school. Upon successful connection, it displays the message "You're a student now!". This script will be useful for establishing SSH connections with the server using a private key.