feat: Update SSH connection script to use a variable for the server IP
This commit updates the 0x0B-ssh/0-use_a_private_key script to use a variable for the server IP address. Instead of directly specifying the IP address in the script, a variable called "server" is introduced and set to "52.86.159.169". This change allows for easier modification of the server IP address in the future. The script still connects to the server using the private key located at ~/.ssh/school and displays the message "You're a student now!". This update improves the flexibility and maintainability of the SSH connection script.