feat: Add script to display the size of the response body
This commit adds a new script, 0-body_size.sh, which uses the curl command to retrieve the size of the response body from a given URL. The script first sends a HEAD request to the URL using the -sI option to suppress the progress meter and display only the headers. It then uses grep and cut commands to extract the value of the "Content-Length" header, which represents the size of the response body. This script will be useful for checking the size of the response body in network-related tasks.